requirements/base.txt
#
# This file is autogenerated by pip-compile with python 3.7
# To update, run:
#
#    pip-compile base.in
#
black==22.10.0
    # via -r base.in
click==8.1.3
    # via
    #   black
    #   flask
flake8==5.0.4
    # via -r base.in
flask==2.2.2
    # via -r base.in
importlib-metadata==4.2.0
    # via
    #   click
    #   flake8
    #   flask
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via flask
markupsafe==2.1.1
    # via
    #   jinja2
    #   werkzeug
marshmallow==3.18.0
    # via webargs
mccabe==0.7.0
    # via flake8
mypy-extensions==0.4.3
    # via black
packaging==21.3
    # via
    #   marshmallow
    #   webargs
pathspec==0.10.1
    # via black
platformdirs==2.5.2
    # via black
psycopg2-binary==2.9.4
    # via -r base.in
pycodestyle==2.9.1
    # via flake8
pyflakes==2.5.0
    # via flake8
pyparsing==3.0.9
    # via packaging
pytz==2022.5
    # via -r base.in
tomli==2.0.1
    # via black
typed-ast==1.5.4
    # via black
typing-extensions==4.4.0
    # via
    #   black
    #   importlib-metadata
webargs==8.2.0
    # via -r base.in
werkzeug==2.2.2
    # via flask
zipp==3.9.0
    # via importlib-metadata
